refactor: canonicalize model paths to models/ (stage 16)
Behavior-preserving. Model definitions moved to models/readwise.py (v2
concepts) and models/reader.py (v3 concepts); v2/models.py and v3/models.py
become identity-preserving re-export shims (verified
v2.models.HighlightCreate is models.readwise.HighlightCreate, and the v3
equivalent). Internal imports across resources/operations/sdk and mcp/server.py
now use canonical paths. Every legacy model import path still resolves to the
same class object. Characterization unchanged; gate green (835 passed);
deps/uv.lock/Justfile untouched.
