+++ src/readwise_sdk/operations/books.py
@@ -10,7 +10,7 @@ from datetime import UTC, datetime, timedelta
 from typing import Protocol
 
 from readwise_sdk.models import BookSearch, BookSummary
-from readwise_sdk.v2.models import Book, BookCategory, Highlight
+from readwise_sdk.models.readwise import Book, BookCategory, Highlight
 
 
 class BooksResource(Protocol):
