+++ src/readwise_sdk/resources/v2/books.py
@@ -6,9 +6,9 @@ from collections.abc import AsyncIterator
 from datetime import datetime
 from typing import Any
 
+from readwise_sdk.models.readwise import Book, BookCategory
 from readwise_sdk.transport.async_ import AsyncTransport
 from readwise_sdk.transport.pagination import StandardV2Page, paginate_async
-from readwise_sdk.v2.models import Book, BookCategory
 
 
 class AsyncBooksResource:
