+++ src/readwise_sdk/resources/v2/export.py
@@ -6,9 +6,9 @@ from collections.abc import AsyncIterator
 from datetime import datetime
 from typing import Any
 
+from readwise_sdk.models.readwise import ExportBook
 from readwise_sdk.transport.async_ import AsyncTransport
 from readwise_sdk.transport.pagination import ExportV2Page, paginate_async
-from readwise_sdk.v2.models import ExportBook
 
 
 class AsyncExportResource:
