+++ src/readwise_sdk/resources/v2/tags.py
@@ -4,9 +4,9 @@ from __future__ import annotations
 
 from collections.abc import AsyncIterator
 
+from readwise_sdk.models.readwise import Tag
 from readwise_sdk.transport.async_ import AsyncTransport
 from readwise_sdk.transport.pagination import StandardV2Page, paginate_async
-from readwise_sdk.v2.models import Tag
 
 
 class AsyncTagsResource:
