+++ src/readwise_sdk/resources/v3/tags.py
@@ -4,9 +4,9 @@ from __future__ import annotations
 
 from collections.abc import AsyncIterator
 
+from readwise_sdk.models.reader import DocumentTag
 from readwise_sdk.transport.async_ import AsyncTransport
 from readwise_sdk.transport.pagination import ReaderV3Page, paginate_async
-from readwise_sdk.v3.models import DocumentTag
 
 
 class AsyncTagsResource:
